+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/endpoint/GetUserEndpointData.java
@@ -4,5 +4,6 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import java.util.List;
 
 public interface GetUserEndpointData {
-    List<ObjectNode> execute(Long userId, Long endpointId, int limit, int offset);
+    List<ObjectNode> execute(Long userId, Long endpointId, Long limit,
+                             Long offset);
 }
