+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/endpoint/GetUserEndpointDataImpl.java
@@ -28,13 +28,15 @@ public class GetUserEndpointDataImpl
     private final GetJsonNodeFromDataHub getJsonNodeFromDataHub;
 
     @Override
-    public List<ObjectNode> execute(Long userId, Long endpointId, int limit, int offset) {
+    public List<ObjectNode> execute(Long userId, Long endpointId, Long limit,
+                                    Long offset) {
         final var enableEndpoint =
                 getEnableEndpointByUserAndEndpointId.execute(userId,
                         endpointId);
         final var endpoint = enableEndpoint.getEndpoint();
         final var rawEndpointData =
-                getJsonNodeFromDataHub.execute(endpoint.getSensorUrl(), limit, offset);
+                getJsonNodeFromDataHub.execute(endpoint.getSensorUrl(), limit,
+                        offset);
 
         return parseWeatherData(rawEndpointData, endpoint,
                 enableEndpoint.getEnableFields());
