+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/endpoint/GetUserEndpointDataWithFields.java
@@ -3,5 +3,5 @@ package pl.edu.agh.apdvbackend.use_cases.endpoint;
 import pl.edu.agh.apdvbackend.controllers.endpoint.body_models.EndpointData;
 
 public interface GetUserEndpointDataWithFields {
-    EndpointData execute(Long userId, Long endpointId, int limit, int offset);
+    EndpointData execute(Long userId, Long endpointId, Long limit, Long offset);
 }
