+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/endpoint/GetUserEndpointDataWithFieldsImpl.java
@@ -18,7 +18,8 @@ public class GetUserEndpointDataWithFieldsImpl
     private final FieldMapper fieldMapper;
 
     @Override
-    public EndpointData execute(Long userId, Long endpointId, int limit, int offset) {
+    public EndpointData execute(Long userId, Long endpointId, Long limit,
+                                Long offset) {
         final var fields = getEnableFields
                 .execute(userId, endpointId);
         final var fieldsWithoutId =
